implement authentication

+++ Domain/Commands/DeleteEventByIdCommand.cs
@@ -1,7 +1,9 @@
 using System.Threading;
 using System.Threading.Tasks;
 
+using Domain.Exceptions;
 using Domain.Interfaces;
+using Domain.Services;
 
 using MediatR;
 
@@ -10,6 +12,7 @@ namespace Domain.Commands
     public class DeleteEventByIdCommand : IRequest<DeleteEventByIdResult>
     {
         public string Id { get; init; }
+        public string UserName { get; init; }
     }
     public class DeleteEventByIdResult
     {
@@ -19,14 +22,21 @@ namespace Domain.Commands
     {
 
         private readonly IRepositoryManager _repositoryManager;
+        private readonly AccessChecker _accessChecker;
 
-        public DeleteEventByIdCommandHandler(IRepositoryManager repositoryManager)
+        public DeleteEventByIdCommandHandler(IRepositoryManager repositoryManager, AccessChecker accessChecker)
         {
             _repositoryManager = repositoryManager;
+            _accessChecker = accessChecker;
         }
 
         public async Task<DeleteEventByIdResult> Handle(DeleteEventByIdCommand request, CancellationToken cancellationToken)
         {
+            bool hasAccess = await _accessChecker.HasAccessAsOwner(request.Id, request.UserName, cancellationToken);
+            if (!hasAccess)
+            {
+                throw new NoAccessException("You do not have access to this event");
+            }
             _ = await _repositoryManager.Event.DeleteEventById(request.Id, cancellationToken);
             await _repositoryManager.SaveAsync(cancellationToken);
             return new DeleteEventByIdResult();
